refactor(ui): flatten `help` and `login` dirs
- both have a single component

- also rename `index.tsx` to `index.ts` since it has no JSX
