+++ ui/src/app/help/help.tsx
@@ -1,8 +1,8 @@
 import {Page} from 'argo-ui';
 import * as React from 'react';
 
-import {uiUrl} from '../../shared/base';
-import {useCollectEvent} from '../../shared/components/use-collect-event';
+import {uiUrl} from '../shared/base';
+import {useCollectEvent} from '../shared/components/use-collect-event';
 
 import './help.scss';
