+++ ui/src/app/help/index.ts
@@ -1,4 +1,4 @@
-import {Help} from './components/help';
+import {Help} from './help';
 
 export default {
     component: Help
