+++ ui/src/app/login/index.ts
@@ -1,4 +1,4 @@
-import {Login} from './components/login';
+import {Login} from './login';
 
 export default {
     component: Login
