+++ ui/src/app/login/login.tsx
@@ -1,8 +1,8 @@
 import {Page} from 'argo-ui';
 import * as React from 'react';
 
-import {uiUrl, uiUrlWithParams} from '../../shared/base';
-import {useCollectEvent} from '../../shared/components/use-collect-event';
+import {uiUrl, uiUrlWithParams} from '../shared/base';
+import {useCollectEvent} from '../shared/components/use-collect-event';
 
 import './login.scss';
